Cleanup Special Event Code
- Remove SpecialEvent namespace
- Separate Events into their own modules
- Add more sanity checks
- Fix wrong tag path for ball in St Pat event
- Rename HaloString to e_global_string_ids
