+++ xlive/H2MOD/Modules/SpecialEvents/Events/Birthday.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void BirthdayOnMapLoad();
