+++ xlive/H2MOD/Modules/SpecialEvents/Events/Christmas.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void ChristmasOnMapLoad();
