+++ xlive/H2MOD/Modules/SpecialEvents/Events/Halloween.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "Blam/Common/Common.h"
+
+void halloween_game_life_cycle_update(e_game_life_cycle state);
+void HalloweenOnMapLoad();
