+++ xlive/H2MOD/Modules/SpecialEvents/Events/Mook.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void MookMaddnessOnMapLoad();
