+++ xlive/H2MOD/Modules/SpecialEvents/Events/Paddy.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void PaddysOnMapLoad();
