+++ xlive/Project_Cartographer.vcxproj
@@ -304,6 +304,12 @@
     <ClCompile Include="H2MOD\Modules\RenderHooks\RenderHooks.cpp" />
     <ClCompile Include="H2MOD\Modules\Shell\Shell.cpp" />
     <ClCompile Include="H2MOD\Modules\Shell\Startup\WinMainH2.cpp" />
+    <ClCompile Include="H2MOD\Modules\SpecialEvents\Events\Birthday.cpp" />
+    <ClCompile Include="H2MOD\Modules\SpecialEvents\Events\Christmas.cpp" />
+    <ClCompile Include="H2MOD\Modules\SpecialEvents\Events\Halloween.cpp" />
+    <ClCompile Include="H2MOD\Modules\SpecialEvents\Events\Mook.cpp" />
+    <ClCompile Include="H2MOD\Modules\SpecialEvents\Events\Paddy.cpp" />
+    <ClCompile Include="H2MOD\Modules\SpecialEvents\SpecialEventHelpers.cpp" />
     <ClCompile Include="H2MOD\Modules\SpecialEvents\SpecialEvents.cpp" />
     <ClCompile Include="H2MOD\Modules\Stats\StatsHandler.cpp" />
     <ClCompile Include="H2MOD\Modules\TagFixes\TagFixes.cpp" />
@@ -627,8 +633,14 @@
     <ClInclude Include="H2MOD\Modules\Shell\ServerConsole.h" />
     <ClInclude Include="H2MOD\Modules\Shell\Shell.h" />
     <ClInclude Include="H2MOD\Modules\Shell\Startup\WinMainH2.h" />
-    <ClInclude Include="H2MOD\Modules\SpecialEvents\MapObjectPlacements\Halloween\Coagulation.h" />
-    <ClInclude Include="H2MOD\Modules\SpecialEvents\MapObjectPlacements\Halloween\Lockout.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\Birthday.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\Christmas.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\Halloween.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\MapObjectPlacements\Halloween\Coagulation.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\MapObjectPlacements\Halloween\Lockout.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\Mook.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\Events\Paddy.h" />
+    <ClInclude Include="H2MOD\Modules\SpecialEvents\SpecialEventHelpers.h" />
     <ClInclude Include="H2MOD\Modules\SpecialEvents\SpecialEvents.h" />
     <ClInclude Include="H2MOD\Modules\Stats\StatsHandler.h" />
     <ClInclude Include="H2MOD\Modules\TagFixes\TagFixes.h" />
